requirements_Auto_BFE.txt
absl-py==0.7.1
affine==2.2.2
altair==2.3.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
branca==0.3.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cloudpickle==1.2.1
cryptography==2.3.1
cycler==0.10.0
cytoolz==0.10.0
dask==2.2.0
decorator==4.3.0
defusedxml==0.6.0
descartes==1.1.0
entrypoints==0.2.3
Fiona==1.8.4
folium==0.7.0
gast==0.2.2
geographiclib==1.49
geojson==2.5.0
geopandas==0.4.0
geopy==1.17.0
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
homura==0.1.5
humanize==0.5.1
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
jsonschema==3.0.0a3
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
kml2geojson==4.0.2
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mercantile==1.0.4
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.1.0
munch==2.3.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.2
notebook==5.7.2
numpy==1.15.4
olefile==0.46
opencv-python==4.1.2.30
osmnx==0.8.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.9.0
psycopg2==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
PyCRS==0.1.3
pycurl==7.43.0.2
Pygments==2.3.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyproj==2.4.2.post1
pyrsistent==0.14.7
PySAL==1.14.4.post2
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==17.1.2
rasterio==1.0.18
rasterstats==0.13.1
requests==2.21.0
Rtree==0.8.3
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplejson==3.16.0
six==1.12.0
sklearn==0.0
snuggs==1.4.2
SQLAlchemy==1.2.14
supermercado==0.0.5
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toml==0.10.0
toolz==0.9.0
torch==1.1.0
torchvision==0.3.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
vincent==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wordcloud==1.6.0
wrapt==1.11.2
